restructuring and removal of deprecated code
