+++ molSimplify/Informatics/deprecated/README
@@ -1,0 +1,1 @@
+Use lacRACAssemble instead of these scripts (autocorrelation and clean_autocorrelation).
