+++ molSimplify/Scripts/deprecated/deprecated_gui_code/grabguivars.py
